Test(PixabayAPI): Finishes unit tests and removes button not working tests
